+++ ximpol/bin/xpobssim.py
@@ -53,10 +53,11 @@ def xpobssim(output_file_path, config_file_path, irf_name, duration, start_time,
     module_name = os.path.basename(config_file_path).replace('.py', '')
     ROI_MODEL = imp.load_source(module_name, config_file_path).ROI_MODEL
     stop_time = start_time + duration
+    gti_list = [(start_time, stop_time)]
     sampling_time = numpy.linspace(start_time, stop_time, time_steps)
     event_list=ROI_MODEL.rvs_event_list( aeff, psf, modf, edisp,sampling_time)
     logger.info('Done %s.' % chrono)
-    event_list.write_fits(output_file_path)
+    event_list.write_fits(output_file_path, gti_list, ROI_MODEL)
     logger.info('All done %s!' % chrono)