+++ doc/release_notes.rst
@@ -3,6 +3,8 @@ Release notes
 
 * Module ximpol.core.fitsio added (issue #49).
 * ximpol.evt.event refactored to use the new ximpol.core.fitsio module.
+* GTI list in the output event file (issue #24)
+* ROI table in the output event file (issue #45).
 
 
 *ximpol (0.15.2) - Mon, 25 Jan 2016 18:04:33 -0800*